+++ activiti-cloud-service-common/activiti-cloud-service-messaging-starter/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+  org.activiti.cloud.common.messaging.config.ActivitiCloudMessagingStarterAutoConfiguration
