+++ activiti-cloud-service-common/activiti-cloud-service-messaging-starter/src/main/resources/config/activiti-cloud-stream-binders-configuration.properties
