--- activiti-cloud-service-common/activiti-cloud-stream-binders-starter/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.activiti.cloud.stream.binders.config.CloudStreamBindersAutoConfiguration
